Move music
Clean up code and fix typos
