+++ resources/lib/objects/__init__.py
@@ -2,3 +2,4 @@
 from movies import Movies
 from musicvideos import MusicVideos
 from tvshows import TVShows
+from music import Music
